+++ src/app/loadout-builder/generated-sets/CompareDrawer.tsx
@@ -8,6 +8,7 @@ import { updateLoadout } from 'app/loadout-drawer/actions';
 import { Loadout, LoadoutItem } from 'app/loadout-drawer/loadout-types';
 import { getModRenderKey } from 'app/loadout/mod-utils';
 import { useD2Definitions } from 'app/manifest/selectors';
+import { UpgradeSpendTier } from 'app/settings/initial-settings';
 import { RootState, ThunkDispatchProp } from 'app/store/types';
 import { DestinyClass } from 'bungie-api-ts/destiny2';
 import clsx from 'clsx';
@@ -18,21 +19,18 @@ import { connect } from 'react-redux';
 import { getItemsFromLoadoutItems } from '../../loadout-drawer/loadout-utils';
 import { assignModsToArmorSet } from '../mod-utils';
 import { getTotalModStatChanges } from '../process/mappers';
-import {
-  ArmorSet,
-  LockableBucketHashes,
-  statHashes,
-  statKeys,
-  StatTypes,
-  UpgradeSpendTier,
-} from '../types';
+import { ArmorSet, LockableBucketHashes, statHashes, statKeys, StatTypes } from '../types';
 import { getPower, upgradeSpendTierToMaxEnergy } from '../utils';
 import styles from './CompareDrawer.m.scss';
 import Mod from './Mod';
 import SetStats from './SetStats';
 import Sockets from './Sockets';
 
-function getItemStats(defs: D2ManifestDefinitions, item: DimItem, upgradeSpendTier: number) {
+function getItemStats(
+  defs: D2ManifestDefinitions,
+  item: DimItem,
+  upgradeSpendTier: UpgradeSpendTier
+) {
   const baseStats = _.mapValues(
     statHashes,
     (value) => item.stats?.find((s) => s.statHash === value)?.base || 0