+++ src/app/loadout-builder/utils.ts
@@ -6,7 +6,20 @@ import { warnLog } from 'app/utils/log';
 import { DestinyEnergyType, DestinyInventoryItemDefinition } from 'bungie-api-ts/destiny2';
 import _ from 'lodash';
 import { ProcessItem } from './process-worker/types';
-import { LockedItemType, UpgradeSpendTier } from './types';
+import { LockedItemType } from './types';
+
+// Todo(ryan): Temporary until api types are available
+// this is a copy of the initial settings one as we can't have a link
+// between the web worker for process and initial settings.
+export enum UpgradeSpendTier {
+  Nothing,
+  LegendaryShards,
+  EnhancementPrisms,
+  AscendantShardsNotExotic,
+  AscendantShards,
+  AscendantShardsNotMasterworked,
+  AscendantShardsLockEnergyType,
+}
 
 /**
  * Add a locked item to the locked item list for a bucket.
@@ -102,9 +115,17 @@ function getEnergySpendTierBoundaryHash(item: DimItem, tier: UpgradeSpendTier) {
       if (!isExotic) {
         break;
       }
+      // for exotics we allow energy upgrades/swaps using enhancement prisms.
       boundaryHash = UpgradeMaterialHashes.ascendantShard;
       break;
     }
+    case UpgradeSpendTier.AscendantShardsNotMasterworked:
+      // already masterworked items will have full energy by default, by dropping the boundary
+      // we will stop energy swaps
+      boundaryHash =
+        item.energy?.energyCapacity === 10 ? UpgradeMaterialHashes.ascendantShard : 'none';
+      break;
+    case UpgradeSpendTier.AscendantShardsLockEnergyType:
     case UpgradeSpendTier.AscendantShards:
       break;
   }
@@ -128,11 +149,13 @@ export function upgradeSpendTierToMaxEnergy(
 
   const boundaryHash = getEnergySpendTierBoundaryHash(item, tier);
 
+  // gets all possible energy upgrades for the item, including the current level
+  // we need this to populate previous item in the loops below
   const availableEnergyUpgrades = energyUpgrade(
     defs,
     item,
     item.energy.energyType,
-    item.energy.energyCapacity - 1, // allows us to get the at level tier.
+    item.energy.energyCapacity - 1,
     item.energy.energyType,
     10
   );
@@ -147,6 +170,8 @@ export function upgradeSpendTierToMaxEnergy(
     );
   }
 
+  // find the upgrade that uses the boundary material, the one before this will be the max
+  // upgrade we can use
   let previousUpgrade: DestinyInventoryItemDefinition | undefined;
   for (const upgrade of availableEnergyUpgrades) {
     const upgradeItem = defs.InventoryItem.get(upgrade);
@@ -154,6 +179,7 @@ export function upgradeSpendTierToMaxEnergy(
       upgradeItem.plug!.insertionMaterialRequirementHash
     );
 
+    // if we find the boundary material we have gone too far
     if (upgradeMaterials.materials.some((material) => material.itemHash === boundaryHash)) {
       break;
     }
@@ -171,7 +197,11 @@ export function canSwapEnergyFromUpgradeSpendTier(
   tier: UpgradeSpendTier,
   item: DimItem
 ) {
-  if (!item.energy || tier === UpgradeSpendTier.Nothing) {
+  if (
+    !item.energy ||
+    tier === UpgradeSpendTier.Nothing ||
+    tier === UpgradeSpendTier.AscendantShardsLockEnergyType
+  ) {
     return false;
   }
 
@@ -199,15 +229,17 @@ export function canSwapEnergyFromUpgradeSpendTier(
     }
   }
 
+  // gets a single upgrade for swapping energy at the current level
   const availableEnergyUpgrades = energyUpgrade(
     defs,
     item,
     item.energy.energyType,
-    item.energy.energyCapacity, // allows us to get the at level tier.
+    item.energy.energyCapacity,
     differentEnergy,
     item.energy.energyCapacity
   );
 
+  // if an energy upgrade is not found it means we cannot swap
   if (!availableEnergyUpgrades.length) {
     return false;
   }
@@ -218,11 +250,13 @@ export function canSwapEnergyFromUpgradeSpendTier(
     return true;
   }
 
-  // There should only be one upgrade possibility due to how energyUpgrade works
   const upgrade = defs.InventoryItem.get(availableEnergyUpgrades[0]);
   const upgradeMaterials = defs.MaterialRequirementSet.get(
     upgrade.plug!.insertionMaterialRequirementHash
   );
 
+  // Make sure that the boundary material is not in the necessary upgrade material
+  // for example a masterworked item will need ascendant shards, enhancement prisms and
+  // legendary shards to swap to another masterwork.
   return upgradeMaterials.materials.every((material) => material.itemHash !== boundaryHash);
 }