+++ docs/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Armor 1 exotics are visible in the exotic picker, albeit unselectable.
 * Default to similar loadout as comparison base in Loadout Optimizer.
+* Armor upgrades in the Optimizer have full descriptions of their functionality. Added Ascendant Shard 'Not Masterworked' and 'Lock Energy Type' options.
+* In the Exotic Selector, the currently selected exotic is now highlighted.
 
 ## 6.70.0 <span class="changelog-date">(2021-06-23)</span>